+++ extensions/hidemyemail/CHANGELOG.md
@@ -1,3 +1,7 @@
 # HideMyEmail Changelog
 
+## [Update] - 2024-12-20
+
+- Supports Chinese Apple ID accounts.
+
 ## [Initial Version] - 2024-12-16